# docs/api.yml
openapi: 3.0.0
info:
  title: UIT-Go Ride-Hailing Platform API
  version: 1.0.0
  description: |
    RESTful API specification for the UIT-Go microservices architecture.
    This specification covers all three core services: UserService, TripService, and DriverService.

    **Base URLs:**
    - UserService: https://api.uitgo.example.com/users
    - TripService: https://api.uitgo.example.com/trips
    - DriverService: https://api.uitgo.example.com/drivers

    **Authentication:**
    All protected endpoints require a JWT Bearer token in the Authorization header.
servers:
  - url: https://api.uitgo.example.com
    description: Production API Gateway (ALB)
  - url: http://localhost:3000
    description: Local development via Docker Compose
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp, requestId]
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Email is required'
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid

    # User Schemas
    UserRole:
      type: string
      enum: [PASSENGER, DRIVER]
    User:
      type: object
      required: [id, email, role, firstName, lastName, phoneNumber, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password, role, firstName, lastName, phoneNumber]
      properties:
        email:
          type: string
          format: email
          example: 'passenger@example.com'
        password:
          type: string
          minLength: 8
          example: 'SecurePass123!'
        role:
          $ref: '#/components/schemas/UserRole'
        firstName:
          type: string
          example: 'John'
        lastName:
          type: string
          example: 'Doe'
        phoneNumber:
          type: string
          example: '+84901234567'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
          description: JWT token for authenticated requests
        user:
          $ref: '#/components/schemas/User'

    # Driver Profile Schemas
...
paths:
...
  /users/register:
    post:
      tags: [UserService]
      summary: Register a new user (passenger or driver)
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/login:
    post:
      tags: [UserService]
      summary: Authenticate user and receive JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...