# packages/shared-types/src/enums/index.ts
/**
 * User Role Enum
 * Defines the role of a user in the system
 */
export enum UserRole {
  PASSENGER = 'PASSENGER',
  DRIVER = 'DRIVER',
}

/**
 * Driver Profile Approval Status
 */
export enum ApprovalStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
}